#10 Renamed package for launchpad s api module
