+++ launchpad-s-api/src/test/kotlin/io/lp4k/launchpad/api/s/LightIntensityLaunchSTest.kt
@@ -14,7 +14,7 @@
  *    limitations under the License.
  *
  */
-package io.lp4k.api
+package io.lp4k.launchpad.api.s
 
 import org.junit.jupiter.api.Assertions.assertDoesNotThrow
 import org.junit.jupiter.api.Assertions.assertEquals
