+++ launchpad-s-api/src/test/kotlin/io/lp4k/launchpad/api/s/ScrollSpeedLaunchSTest.kt
@@ -14,9 +14,9 @@
  *    limitations under the License.
  *
  */
-package io.lp4k.api
+package io.lp4k.launchpad.api.s
 
-import io.lp4k.api.ScrollSpeedLaunchS.Companion.of
+import io.lp4k.launchpad.api.s.ScrollSpeedLaunchS.Companion.of
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotEquals
 import org.junit.jupiter.api.Test
