+++ lp4k-emu-web/src/test/kotlin/io/lp4k/emulator/EmulatorLaunchpadClientTest.kt
@@ -16,10 +16,10 @@
  */
 package io.lp4k.emulator
 
-import io.lp4k.api.*
 import io.lp4k.emulator.output.OutputEventType
 import io.lp4k.launchpad.api.BackBufferOperation
 import io.lp4k.launchpad.api.Buffer
+import io.lp4k.launchpad.api.s.*
 import io.vertx.core.json.JsonObject
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
