+++ lp4k-midi/src/main/kotlin/io/lp4k/midi/protocol/DefaultMidiProtocolListener.kt
@@ -16,8 +16,8 @@
  */
 package io.lp4k.midi.protocol
 
-import io.lp4k.api.ButtonLaunchS
-import io.lp4k.api.PadLaunchS
+import io.lp4k.launchpad.api.s.ButtonLaunchS
+import io.lp4k.launchpad.api.s.PadLaunchS
 import io.lp4k.launchpad.api.LaunchpadListener
