+++ lp4k-midi/src/test/kotlin/io/lp4k/midi/protocol/DefaultMidiProtocolListenerTest.kt
@@ -17,8 +17,8 @@
  */
 package io.lp4k.midi.protocol
 
-import io.lp4k.api.ButtonLaunchS
-import io.lp4k.api.PadLaunchS
+import io.lp4k.launchpad.api.s.ButtonLaunchS
+import io.lp4k.launchpad.api.s.PadLaunchS
 import io.lp4k.launchpad.api.LaunchpadListener
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
